Organisation des fichiers selon les exigences du projet
