+++ Authentification/src/main.py
@@ -1,0 +1,11 @@
+import os
+
+from dotenv import load_dotenv
+
+from Authentification.src import app
+
+load_dotenv()
+
+if __name__ == "__main__":
+    debug_mode = os.getenv("FLASK_DEBUG", "false").lower() == "true"
+    app.run(debug=debug_mode)
