+++ Authentification/src/models/example_model.py
@@ -1,0 +1,21 @@
+from ..routes import db
+
+
+class User(db.Model):
+    __tablename__ = 'users'
+
+    nom = db.Column(db.String(50), nullable=False)
+    prenom = db.Column(db.String(50), nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    password = db.Column(db.String(120), nullable=False)
+    username = db.Column(db.String(50), unique=True, nullable=False)
+    pseudo = db.Column(db.String(50), nullable=False)
+
+    def to_dict(self):
+        return {
+            "nom": self.nom,
+            "prenom": self.prenom,
+            "email": self.email,
+            "username": self.username,
+            "pseudo": self.pseudo
+        }
