+++ Authentification/src/routes.py
@@ -1,25 +1,7 @@
-from flask import Flask, request, jsonify
-from flask_jwt_extended import JWTManager
+from flask import request, jsonify
 from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
-from flask_sqlalchemy import SQLAlchemy
-from datetime import timedelta
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-app = Flask(__name__)
-
-app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv("DATABASE_URL", "sqlite:///app.db")
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['JWT_SECRET_KEY'] = os.getenv("JWT_SECRET_KEY", "super-secret")
-app.config['JWT_ACCESS_TOKEN_EXPIRES'] = timedelta(minutes=15)
-app.config['JWT_REFRESH_TOKEN_EXPIRES'] = timedelta(days=1)
-app.config['JWT_TOKEN_LOCATION'] = ['headers']
-app.config['JWT_COOKIE_CSRF_PROTECT'] = False
-
-db = SQLAlchemy(app)
-jwt = JWTManager(app)
+from Authentification.src import app, db
+from Authentification.src.models.example_model import User
 
 
 @jwt_required()
@@ -27,26 +9,6 @@ def refresh_token_if_needed():
     current_user = get_jwt_identity()
     new_token = create_access_token(identity=current_user)
     return jsonify({"new_token": new_token}), 200
-
-
-class User(db.Model):
-    __tablename__ = 'users'
-
-    nom = db.Column(db.String(50), nullable=False)
-    prenom = db.Column(db.String(50), nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    password = db.Column(db.String(120), nullable=False)
-    username = db.Column(db.String(50), unique=True, nullable=False)
-    pseudo = db.Column(db.String(50), nullable=False)
-
-    def to_dict(self):
-        return {
-            "nom": self.nom,
-            "prenom": self.prenom,
-            "email": self.email,
-            "username": self.username,
-            "pseudo": self.pseudo
-        }
 
 
 # Endpoint pour l'incription
@@ -152,8 +114,3 @@ def update_user(username):
 
     except Exception as e:
         return jsonify({"message": "Erreur interne du serveur", "error": str(e)}), 500
-
-
-if __name__ == "__main__":
-    debug_mode = os.getenv("FLASK_DEBUG", "false").lower() == "true"
-    app.run(debug=debug_mode)
